requirements.txt
pygame>=2.0.0
PyOpenGL>=3.1.0
numpy>=1.19.0
pyrr>=0.10.3